securitycenter: refactoring to use the new pattern
